Build the genesis directly in the repository
